infra/paddleocr/app/requirements.txt
# FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PaddleOCR
paddleocr==2.7.0.3
paddlepaddle==2.5.1

# Image processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1

# Monitoring
prometheus-client==0.19.0

# Logging
python-json-logger==2.0.7
